moved contracts to fpi and adding contracts config
